requirement.txt
djangorestframework
django-cors-headers
requests
beautifulsoup4
